move moveSubtree to its own module to reduce Vite's screaming
